requirements.txt
# Web框架和服务器
fastapi==0.115.14
uvicorn==0.35.0
starlette==0.46.2
pydantic==2.11.7
pydantic-settings==2.10.1

# HTTP客户端和网络
httpx==0.28.1
aiohttp==3.12.13
requests==2.32.4
websockets==15.0.1

# LangChain生态系统 - 核心框架
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langsmith==0.4.4

# 向量数据库和存储
pymilvus==2.5.12
langchain-milvus==0.2.1
milvus-lite==2.5.1
redis==6.2.0
qdrant-client==1.14.3

# 数据库连接
psycopg2-binary==2.9.10
asyncpg==0.30.0
SQLAlchemy==2.0.41

# 文档处理和加载器
unstructured==0.18.11
pypdf==5.9.0
python-docx==1.2.0
beautifulsoup4==4.13.4
pdfplumber==0.11.7
lxml==6.0.0
Markdown==3.8.2

# AI/ML模型和服务
openai==1.93.0
anthropic==0.57.1
dashscope==1.23.6
sentence-transformers==5.0.0
transformers==4.53.1
torch==2.7.1
tokenizers==0.21.2
tiktoken==0.9.0

# 科学计算和数据处理
numpy==2.3.1
pandas==2.3.0
scikit-learn==1.7.0
scipy==1.16.0

# 自然语言处理
nltk==3.9.1
langdetect==1.0.9

# 异步和并发处理
aiofiles==24.1.0
asyncio  # 内置模块，不需要版本
nest-asyncio==1.6.0
celery==5.5.3

# 配置和环境管理
python-dotenv==1.1.1
PyYAML==6.0.2
click==8.2.1

# 数据验证和序列化
marshmallow==3.26.1
orjson==3.10.18

# 安全和认证
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20

# 监控和日志
prometheus-client==0.22.1
structlog==25.4.0

# 测试框架
pytest==8.4.1
pytest-asyncio==1.0.0

# 代码质量工具
black==25.1.0
isort==6.0.1
mypy==1.16.1

# 工具库
tqdm==4.67.1
tenacity==9.1.2
backoff==2.2.1
packaging==24.2
filelock==3.15.4
fsspec==2025.5.1
joblib==1.5.1

# 开发和调试工具
watchfiles==1.1.0
rich  # 如果使用的话，需要添加

# 其他依赖
emoji==2.14.1
python-dateutil==2.9.0.post0
six==1.17.0
certifi==2025.6.15
urllib3==2.5.0
charset-normalizer==3.4.2
idna==3.10